+++ SourceCode/Services/IModuleService.cs
@@ -10,6 +10,7 @@ namespace ModulesRegistry.Services
 {
     public interface IModuleService : IDataService<Module>
     {
+        Task<(int Count, string Message)> CloneAsync(ClaimsPrincipal? principal, int id, int owningPersonId);
         Task<Module?> FindByIdAsync(ClaimsPrincipal? principal, int id, int owningPersionId);
         Task<IEnumerable<Module>> GetAllAsync(ClaimsPrincipal? principal);
         Task<IEnumerable<Module>> GetForOwningPerson(ClaimsPrincipal? principal, int personId);